[TASK] Cleanup dualprocshm
- Reorganize target specific headers to avoid
  compilation error in Linux kernel.
- Fix some function declaration and documentation.

+++ contrib/dualprocshm/src/dualprocshm-noos.c
@@ -56,9 +56,6 @@ SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 //------------------------------------------------------------------------------
 #include <dualprocshm.h>
 
-#include <stdlib.h>
-#include <string.h>
-
 //============================================================================//
 //            G L O B A L   D E F I N I T I O N S                             //
 //============================================================================//
@@ -514,13 +511,13 @@ tDualprocReturn dualprocshm_readData(tDualprocDrvInstance pInstance_p, UINT8 id_
 {
     tDualProcDrv*   pDrvInst = (tDualProcDrv*)pInstance_p;
     UINT8*          pBase;
-    UINT32          highAddr;
+    size_t          highAddr;
 
     if (pInstance_p == NULL ||  id_p > MAX_DYNAMIC_BUFF_COUNT || pData_p == NULL)
         return kDualprocInvalidParameter;
 
     pBase = pDrvInst->pDynResTbl[id_p].pBase;
-    highAddr = (UINT32)(pBase + pDrvInst->pDynResTbl[id_p].pMemInst->span);
+    highAddr = (size_t)(pBase + pDrvInst->pDynResTbl[id_p].pMemInst->span);
 
     if ((offset_p + size_p) > highAddr)
         return kDualprocNoResource;
@@ -555,13 +552,13 @@ tDualprocReturn dualprocshm_writeData(tDualprocDrvInstance pInstance_p, UINT8 id
 {
     tDualProcDrv*   pDrvInst = (tDualProcDrv*)pInstance_p;
     UINT8*          pBase;
-    UINT32          highAddr;
+    size_t          highAddr;
 
     if (pInstance_p == NULL ||  id_p > MAX_DYNAMIC_BUFF_COUNT || pData_p == NULL)
         return kDualprocInvalidParameter;
 
     pBase = pDrvInst->pDynResTbl[id_p].pBase;
-    highAddr = (UINT32)(pBase + pDrvInst->pDynResTbl[id_p].pMemInst->span);
+    highAddr = (size_t)(pBase + pDrvInst->pDynResTbl[id_p].pMemInst->span);
 
     if ((offset_p + size_p) > highAddr)
         return kDualprocNoResource;
